refactor client source: seperate source file for TUI
