+++ CMakeLists.txt
@@ -3,10 +3,13 @@ cmake_minimum_required(VERSION 3.5)
 project(updog)
 include_directories(include)
 
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+
 find_package(Curses REQUIRED)
 include_directories(${CURSES_INCLUDE_DIR})
 
-add_executable(updog src/utils.c  src/client.c src/main.c src/server.c)
+FILE(GLOB_RECURSE sources src/*.c)
+add_executable(updog ${sources})
 
 target_include_directories(updog PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
